+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./condition/index.js";
 export * from "./predicate/index.js";
+export * from "./typeof/index.js";