+++ test/unit/suite.spec.ts
@@ -1,8 +1,10 @@
 import { suite } from "node:test";
 import { conditionTests } from "./condition/index.js";
 import { predicateTests } from "./predicate/index.js";
+import { typeofTests } from "./typeof/index.js";
 
 suite("suite", () => {
   conditionTests();
   predicateTests();
+  typeofTests();
 });